ARIL/requirements.txt
streamlit
PyMuPDF>=1.24.0
sentence-transformers
chromadb
google-genai
pillow
